pyproject.toml
[tool.poetry]
name = "ocr-agl"
version = "1.1.0"
description = "OCR API: This OCR API is an application for extracting text from images and PDF files. It is built using Flask, a Python web framework. It utilizes the pytesseract OCR library, pymupdf and the PIL library for image processing."
authors = ["Max Hild <Maxhild10@gmail.com>", "Thomas Lux <lux_t1@ukw.de>"]
license = "Apache License 2.0"
readme = "README.md"
packages = [{include = "ocr_agl"}]

[tool.poetry.dependencies]
python = "^3.10.8"
anyio = "==3.7.0"
blinker = "==1.6.2"
certifi = "==2023.5.7"
cffi = "==1.15.1"
charset-normalizer = "==3.1.0"
ci-info = "==0.3.0"
flask = "==2.3.2"
jsonify = "==0.5"
numpy = "==1.24.3"
pandas = "==2.0.2"
pdf2image = "==1.16.3"
pillow = "==9.5.0"
poppler-utils = "==0.1.0"
prov = "==2.0.0"
pycparser = "==2.21"
pydot = "==1.4.2"
pymupdf = "==1.22.3"
pyparsing = "==3.0.9"
pytesseract = "==0.3.10"
requests = "==2.31.0"
simplejson = "==3.19.1"
six = "==1.16.0"
sniffio = "==1.3.0"
starlette = "==0.28.0"
tools = "==0.1.9"
traits = "==6.3.2"
tzdata = "==2023.3"
urllib3 = "==2.0.3"
uvicorn = "==0.22.0"
werkzeug = "==2.3.6"
flask-testing = "==0.8.1"
spacy = "==3.6.0"
opencv-python = "==4.8.0.74"
hunspell = "==4.8.0.74"
bs4="0.0.1"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"



[classifiers]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache License 2.0"
]

[tool.poetry.urls]
"Homepage" = "https://github.com/wg-lux/OCR_Reader/tree/main"
"Bug Tracker" = "https://github.com/wg-lux/OCR_Reader/issues"
